+++ src/session.rs
@@ -254,10 +254,10 @@ impl WsGameSession {
             }
             "/bet" => {
                 if v.len() == 2 {
-                    let amount = v[1].to_owned(); // parse as f64
+                    let amount = v[1].to_owned();
                     self.hub_addr.do_send(messages::PlayerActionMessage {
 			id: self.id,
-			player_action: PlayerAction::Bet(amount.parse::<f64>().unwrap()),
+			player_action: PlayerAction::Bet(amount.parse::<u32>().unwrap()),
                     });
                     ctx.text(format!("placing bet of: {:?}", v[1]));
                 } else {
